% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_similaridade.R
\name{cluster}
\alias{cluster}
\title{Title}
\usage{
cluster(consultoria, datatype, endpoint)
}
\arguments{
\item{method}{seleciona o método do pacote vegan::vegdist, podendo ser:"manhattan", "euclidean", "canberra", "clark",}

\item{ylab}{}
}
\description{
Title
}
