% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curva_coletor.R
\name{coletor}
\alias{coletor}
\title{Curva de acúmulo de espécies}
\usage{
coletor(x, campanha, especie, at, labels)
}
\arguments{
\item{x}{nome do data frame que será utilizado no arquivo}

\item{campanha}{coluna com as campanhas coletadas}

\item{especie}{coluna com a identificacao de todos os individuos coletados}

\item{at}{número de linhas (campanha) onde foram coletadas}

\item{labels}{especificar, em forma de vetor, as campanhas coletadas}
}
\value{
um gráfico
}
\description{
Curva de acúmulo de espécies
}
\examples{
\dontrun{
coletor(consultoria, campanha, especie, at=c(1:4), labels=c("Primeira", "Segunda", "Terceira", "Quarta"))
}
}
