% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabundancia.R
\name{gabundancia}
\alias{gabundancia}
\title{Gera um gráfico de abundância de espécies}
\usage{
gabundancia(
  x,
  campanha,
  especie,
  fill = "#ADC6CB",
  filtro_campanha = NULL,
  theme_few,
  axis.text,
  axis.text.x,
  axis.title
)
}
\arguments{
\item{x}{nome da tabela que será utilizada}

\item{campanha}{coluna que designa o indicador ordinal referente a campanha}

\item{especie}{coluna com a identificacao de todos os individuos coletados}

\item{filtro_campanha}{seleciona a campanha de consultoria, quando não selecionada,
inclui todas}
}
\value{
um gráfico
}
\description{
Gera um gráfico de abundância de espécies
}
\examples{
gabundancia(consultoria, filtro_campanha = "primeira")

}
