% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riqueza.R
\name{riqueza}
\alias{riqueza}
\title{Gera uma tabela com os indices de riqueza (Margalef, Menhinick)
e diversidade (Shannon, Simpson, Pielou)}
\usage{
riqueza(consultoria)
}
\arguments{
\item{consultoria}{nome da tabela que será utilizada}
}
\value{
uma tabela em excel
}
\description{
Gera uma tabela com os indices de riqueza (Margalef, Menhinick)
e diversidade (Shannon, Simpson, Pielou)
}
\examples{
riqueza()
}
