% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_similaridade.R
\name{similaridade}
\alias{similaridade}
\title{Gera um gráfico de similaridade com o índice de Bray-Curtis}
\usage{
similaridade(
  x,
  localidade = localidade,
  especie = especie,
  method = "bray",
  method2 = "average",
  labels = c("Chiador", "Duas Barras", "Macaé", "Nova Friburgo", "Sumidouro",
    "Trajano de Moraes"),
  ylab = "Similaridade (índice de Bray-Curtis)",
  xlab,
  sub
)
}
\arguments{
\item{x}{nome do data frame que será utilizado no arquivo}

\item{localidade}{coluna que designa os locais onde os exemplares foram coletados}

\item{especie}{coluna com a identificação de todos os indivíduos coletados}

\item{method}{seleciona o método do pacote vegan::vegdist, podendo ser:"manhattan", "euclidean", "canberra", "clark",}

\item{labels}{descreve as legendas das localidades que serão as identificações dos terminais dos gráficos}

\item{ylab}{}
}
\value{
um gráfico
}
\description{
Gera um gráfico de similaridade com o índice de Bray-Curtis
}
\examples{
similaridade(consultoria)
}
