% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabundancia.R
\name{tabundancia}
\alias{tabundancia}
\title{Gera uma tabela de abundância de espécies}
\usage{
tabundancia(x, campanha = campanha, especie = especie)
}
\arguments{
\item{x}{nome do data frame que será utilizado no arquivo}

\item{campanha}{coluna que designa o indicador ordinal referente a campanha}

\item{especie}{coluna com a identificacao de todos os individuos coletados}
}
\value{
uma tabela em excel
}
\description{
Gera uma tabela de abundância de espécies
}
\examples{
\dontrun{
tabundancia(consultoria)
}
}
