% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperatura.R
\name{temperatura}
\alias{temperatura}
\title{Gera um gráfico com duas colunas (uma de número de indivíduos e uma
de temperatura média) por localidade}
\usage{
temperatura(x, localidade, especie, temperatura_c)
}
\arguments{
\item{x}{nome do data frame que será utilizado no arquivo}

\item{localidade}{coluna que designa os locais onde os exemplares
foram coletados e serão agrupados por essa coluna}

\item{especie}{coluna com a identificação de todos os indivíduos coletados}

\item{temperatura_c}{coluna com a temperatura dos indivíduos coletados}
}
\value{
um gráfico de número de espécies e temperatura
}
\description{
Gera um gráfico com duas colunas (uma de número de indivíduos e uma
de temperatura média) por localidade
}
\examples{
temperatura(consultoria)
}
