% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefacao.R
\name{coletor}
\alias{coletor}
\title{Curva de acúmulo de espécies}
\usage{
coletor(x, localidade, especie, at, labels)
}
\arguments{
\item{endpoint}{}
}
\description{
Curva de acúmulo de espécies
}
\examples{
acumulo(consultoria, at=c(1:6), labels=c("Chiador", "Duas Barras", "Macaé", "Nova Friburgo", "Sumidouro", "Trajano de Moraes"))
}
