% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabundancia.R
\name{gabundancia}
\alias{gabundancia}
\title{Gera um gráfico de abundância de espécies}
\usage{
gabundancia(
  x,
  campanha = campanha,
  especie = especie,
  fill = "#ADC6CB",
  filtro_campanha = NULL,
  theme_few = ggthemes::theme_few(),
  axis.text = ggplot2::element_text(size = 6, family = "Arial"),
  axis.text.x = ggplot2::element_text(angle = 90, face = "italic"),
  axis.title = ggplot2::element_text(size = 8)
)
}
\arguments{
\item{x}{nome do data frame que será utilizado no arquivo}

\item{campanha}{coluna que designa o indicador ordinal referente a campanha}

\item{especie}{coluna com a identificação de todos os indivíduos coletados}

\item{filtro_campanha}{seleciona a campanha de consultoria, quando não selecionada,
inclui todas}
}
\value{
um gráfico
}
\description{
Gera um gráfico de abundância de espécies
}
\examples{
\dontrun{
gabundancia(consultoria, filtro_campanha = "primeira")
}
}
