% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabundancia.R
\name{gabundancia}
\alias{gabundancia}
\title{Gera um gráfico de abundância de espécies}
\usage{
gabundancia(consultoria, effort)
}
\arguments{
\item{consultoria}{nome da tabela que será utilizada}

\item{campanha}{seleciona a campanha de consultoria, quando não selecionada,
inclui todas}
}
\value{
um gráfico
}
\description{
Gera um gráfico de abundância de espécies
}
\examples{
gabundancia()
}
