% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabundancia.R
\name{tabundancia}
\alias{tabundancia}
\title{Gera uma tabela de abundância de espécies}
\usage{
tabundancia(x, filtro_campanha)
}
\arguments{
\item{consultoria}{nome da tabela que será utilizada}

\item{campanha}{seleciona a campanha de consultoria, quando não selecionada,
inclui todas}
}
\value{
uma tabela em excel
}
\description{
Gera uma tabela de abundância de espécies
}
\examples{
tabundancia()
}
