% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperatura.R
\name{temperatura}
\alias{temperatura}
\title{Gera um gráfico com duas colunas (uma de número de indivíduos e uma
de temperatura média) por localidade}
\usage{
temperatura(consultoria, campanha)
}
\arguments{
\item{consultoria}{nome da tabela que será utilizada}

\item{campanha}{seleciona a campanha de consultoria, quando não selecionada,
inclui todas}
}
\value{
um gráfico de número de espécies e temperatura
}
\description{
Gera um gráfico com duas colunas (uma de número de indivíduos e uma
de temperatura média) por localidade
}
\examples{
temperatura()
}
